+++ apps/web/dms/src/components/pages/home/index.ts
@@ -1,1 +1,1 @@
-export { HomeDashboardPage } from './DashboardPage';
+export { HomeDashboardPage } from './HomeDashboardPage';